+++ conda/exceptions.py
@@ -8,9 +8,10 @@ from conda._vendor.auxlib.entity import EntityEncoder
 from logging import getLogger
 from traceback import format_exc
 
-from . import CondaError, text_type
+from . import CondaError, text_type, CondaExitZero
 from .compat import iteritems, iterkeys
 log = logging.getLogger(__name__)
+
 
 class LockError(CondaError, RuntimeError):
     def __init__(self, message):
@@ -111,19 +112,19 @@ class CondaEnvironmentError(CondaError, EnvironmentError):
         super(CondaEnvironmentError, self).__init__(msg, *args)
 
 
-class DryRunExit(CondaError):
-    def __init__(self, message):
-        msg = 'Dry run exiting: %s' % message
+class DryRunExit(CondaExitZero):
+    def __init__(self):
+        msg = 'Dry run exiting'
         super(DryRunExit, self).__init__(msg)
 
 
-class CondaSystemExit(CondaError, SystemExit):
+class CondaSystemExit(CondaExitZero, SystemExit):
     def __init__(self, *args):
         msg = ' '.join(text_type(arg) for arg in self.args)
         super(CondaSystemExit, self).__init__(msg)
 
 
-class SubprocessExit(CondaError):
+class SubprocessExit(CondaExitZero):
     def __init__(self, *args, **kwargs):
         super(SubprocessExit, self).__init__(*args, **kwargs)
 
@@ -474,6 +475,8 @@ def conda_exception_handler(func, *args, **kwargs):
         return_value = func(*args, **kwargs)
         if isinstance(return_value, int):
             return return_value
+    except CondaExitZero:
+        return 0
     except CondaRuntimeError as e:
         print_unexpected_error_message(e)
         delete_lock()