+++ conda/install.py
@@ -260,8 +260,12 @@ def binary_replace(data, a, b):
     replaced with `b` and the remaining string is padded with null characters.
     All input arguments are expected to be bytes objects.
     """
-    if on_win and has_pyzzer_entry_point(data):
-        return replace_pyzzer_entry_point_shebang(data, a, b)
+    if on_win:
+        if has_pyzzer_entry_point(data):
+            return replace_pyzzer_entry_point_shebang(data, a, b)
+        # currently we should skip replacement on Windows for things we don't understand.
+        else:
+            return data
 
     def replace(match):
         occurances = match.group().count(a)
@@ -1047,7 +1051,8 @@ def messages(prefix):
     path = join(prefix, '.messages.txt')
     try:
         with open(path) as fi:
-            sys.stdout.write(fi.read())
+            fh = sys.stderr if context.json else sys.stdout
+            fh.write(fi.read())
     except IOError:
         pass
     finally: